requirements.txt
google-api-python-client>=2.170.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
matplotlib>=3.9.0
pytest>=8.3.0
streamlit>=1.38.0
fastapi>=0.115.0
uvicorn>=0.30.0
cryptography>=43.0.0
